+++ packages/antwerp-ui/react-components/src/lib/molecules/datepicker/views/MonthsView.tsx
@@ -1,13 +1,10 @@
 import { format, getDate, getMonth, getYear } from 'date-fns';
 import { classNames } from '../../../../utils/dom.utils';
 import { MonthsViewProps } from '../Datepicker.types';
-import { DEFAULT_LOCALE } from '../../../../constants/settings';
-import { useLocale } from '../../../../utils/time.utils';
 import { titleize } from '../../../../utils/string.utils';
+import { DEFAULT_LOCALE } from '../../../../constants/settings';
 
 export function MonthsView({ onChange, value, activeYear, locale = DEFAULT_LOCALE }: MonthsViewProps) {
-  const dateFnsLocale = useLocale(locale);
-
   const renderMonth = (month: number) => {
     const classes = classNames({
       'is-current': month === getMonth(new Date()) && activeYear === getYear(new Date()),
@@ -17,7 +14,7 @@ export function MonthsView({ onChange, value, activeYear, locale = DEFAULT_LOCAL
     return (
       <td>
         <button type="button" className={classes} onClick={() => onChange(monthValue)}>
-          {titleize(format(monthValue, 'MMMM', dateFnsLocale))}
+          {titleize(format(monthValue, 'MMMM', { locale }))}
         </button>
       </td>
     );